service/pyproject.toml
[project]
name = "aiqfome"
version = "0.1.0"
description = ""
authors = [
    {name = "Anderson Freitas",email = "freitas.dev@proton.me"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref (==3.8.1)",
    "click (==8.1.8)",
    "django (==5.2)",
    "django-simple-history (==3.8.0)",
    "django-mkdocs (==0.1.2)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-simplejwt (==5.5.0)",
    "ghp-import (==2.1.0)",
    "jinja2 (==3.1.6)",
    "markdown (==3.7)",
    "markupsafe (==3.0.2)",
    "materialdash (==0.0.24)",
    "mergedeep (==1.3.4)",
    "mkdocs (==1.6.1)",
    "mkdocs-get-deps (==0.2.0)",
    "packaging (==24.2)",
    "pathspec (==0.12.1)",
    "platformdirs (==4.3.7)",
    "psycopg2-binary (==2.9.10)",
    "pyjwt (==2.9.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "pyyaml (==6.0.2)",
    "pyyaml-env-tag (==0.1)",
    "rest-framework-simplejwt (==0.0.2)",
    "six (==1.17.0)",
    "sqlparse (==0.5.3)",
    "watchdog (==6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "uuid_v7 (==1.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"

[tool.isort]
profile = "black"
